+++ crates/module-system/module-implementations/integration-tests/tests/stf_blueprint/operator/auth_eip712.rs
@@ -1,5 +1,7 @@
 use sov_address::{EthereumAddress, EvmCryptoSpec};
 use sov_evm::Eip712Authenticator;
+use sov_evm::Eip712AuthenticatorInput;
+use sov_evm::Eip712AuthenticatorTrait;
 use sov_evm::SchemaProvider;
 use sov_mock_da::{MockBlob, MockDaSpec};
 use sov_mock_zkvm::MockZkvm;
@@ -91,6 +93,13 @@ generate_runtime! {
     auth_type: Eip712Authenticator<S, TestRuntime<S>, TestSchemaProvider>,
     auth_call_wrapper: |call| call,
 }
+
+impl Eip712AuthenticatorTrait<S> for TestRuntime<S> {
+    fn add_eip712_auth(tx: RawTx) -> <Self::Auth as TransactionAuthenticator<S>>::Input {
+        Eip712AuthenticatorInput::Eip712(tx)
+    }
+}
+
 type RT = TestRuntime<S>;
 
 fn setup() -> (TestRunner<RT, S>, TestUser<S>) {
@@ -132,12 +141,12 @@ pub fn sign_utx<S: Spec, RT: Runtime<S>>(
         .unwrap();
 
     let utx_bytes = borsh::to_vec(&utx).expect("Failed to serialize unsigned transaction");
-    let eip712_hash = schema
-        .eip712_signing_hash(transaction_type_index, &utx_bytes)
+    let eip712_signing_data = schema
+        .eip712_signing_digest(transaction_type_index, &utx_bytes)
         .expect("Failed to calculate EIP712 hash");
 
     let pk = signer.private_key();
-    let signature = pk.sign(&eip712_hash);
+    let signature = pk.sign(&eip712_signing_data);
     utx.to_signed_tx(pk.pub_key(), signature)
 }
 
@@ -157,9 +166,11 @@ pub fn encode_message<S: Spec, RT: Runtime<S> + EncodeCall<ValueSetter<S>>>() ->
     RT::to_decodable(msg)
 }
 
-pub fn encode<S: Spec, RT: Runtime<S>>(tx: Transaction<RT, S>) -> FullyBakedTx {
+pub fn encode<S: Spec, RT: Runtime<S> + Eip712AuthenticatorTrait<S>>(
+    tx: Transaction<RT, S>,
+) -> FullyBakedTx {
     let raw_tx = RawTx::new(borsh::to_vec(&tx).unwrap());
-    RT::Auth::encode_with_standard_auth(raw_tx)
+    <RT as Eip712AuthenticatorTrait<S>>::encode_with_eip712_auth(raw_tx)
 }
 
 fn execute_tx(