+++ crates/module-system/module-implementations/sov-evm/src/rpc/mod.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 
 use alloy_consensus::{transaction::Recovered, Transaction as TransactionTrait, TxReceipt};
-use alloy_primitives::{Address, U64};
+use alloy_primitives::{Address, BlockHash, U64};
 use alloy_primitives::{Bytes, TxKind, B256, U256};
 use alloy_rpc_types::{
     state::StateOverride, Block, BlockOverrides, BlockTransactions, FeeHistory, Log,
@@ -28,9 +28,7 @@ use tracing::debug;
 use crate::conversions::replay_tx_env;
 use crate::db::EvmDb;
 use crate::evm::executor;
-use crate::evm::primitive_types::{
-    Receipt, SealedBlock, TransactionSigned, TransactionSignedAndRecovered,
-};
+use crate::evm::primitive_types::{Receipt, TransactionSigned, TxSignedAndRecovered};
 use crate::executor::{get_cfg_env, inspect, transact_commit};
 use crate::helpers::{
     from_primitive_with_hash, from_recovered_with_block_context, prepare_call_env,
@@ -108,29 +106,30 @@ where
 
         let maybe_block = || -> Option<Block> {
             let block = self.get_sealed_block_by_number(block_number, state)?;
-            let header = from_primitive_with_hash(block.header.clone());
+
+            let block_hash = block.hash().unwrap_or(BlockHash::ZERO);
+            let header = from_primitive_with_hash(block.header().clone(), block_hash);
+
+            let block_number = block.number();
+            let tx_range = block.transactions_start()..block.transactions_end();
 
             let transactions = if Some(true) == details {
                 BlockTransactions::Full(
-                    block
-                        .transactions
-                        .clone()
+                    tx_range
                         .map(|index| {
                             let tx = self.transactions.get(&index, state).unwrap_infallible()?;
                             Some(from_recovered_with_block_context(
                                 tx.into(),
-                                Some(block.header.seal()),
-                                block.header.number,
-                                U256::from(index - block.transactions.start),
+                                Some(block_hash),
+                                block_number,
+                                U256::from(index - block.transactions_start()),
                             ))
                         })
                         .collect::<Option<Vec<_>>>()?,
                 )
             } else {
                 BlockTransactions::Hashes(
-                    block
-                        .transactions
-                        .clone()
+                    tx_range
                         .map(|index| {
                             let tx = self.transactions.get(&index, state).unwrap_infallible()?;
                             Some(*tx.signed_transaction.hash())
@@ -226,7 +225,6 @@ where
         state: &mut ApiStateAccessor<S>,
     ) -> RpcResult<Bytes> {
         debug!("EVM module JSON-RPC request to `eth_getCode`");
-
         let mut state = self.resolve_state(block_number, state)?;
         let code = self
             .accounts
@@ -237,9 +235,10 @@ where
                     .get(&account.code_hash, state.deref_mut())
                     .unwrap_infallible()
             })
+            .map(|code| code.bytecode().clone())
             .unwrap_or_default();
 
-        Ok(code.bytecode().clone())
+        Ok(code.clone())
     }
 
     /// Handler for: `eth_feeHistory`
@@ -269,7 +268,7 @@ where
         let mut maybe_tx = || -> Option<Transaction> {
             let tx_number = self.get_tx_index_by_hash(&hash, state)?;
             let tx = self.transaction(tx_number, state)?;
-            let block = self.get_maybe_sealed_block(tx.block_number, state);
+            let block = self.get_maybe_sealed_block(tx.block_number, state)?;
 
             Some(from_recovered_with_block_context(
                 tx.into(),
@@ -304,7 +303,7 @@ where
         let mut maybe_receipt = || -> Option<TransactionReceipt> {
             let number = self.get_tx_index_by_hash(&hash, state)?;
             let tx = self.transaction(number, state)?;
-            let block = self.get_maybe_sealed_block(tx.block_number, state);
+            let block = self.get_maybe_sealed_block(tx.block_number, state)?;
             let receipt = self.receipt(number, state)?;
             Some(build_rpc_receipt(block, tx, number, receipt))
         };
@@ -353,18 +352,16 @@ where
         debug!("EVM module JSON-RPC request to `eth_estimateGas`");
         let result = self.call(request, block_number, state)?;
         let gas_used = result.gas_used();
-        state
+        let gas_meter = state.try_as_basic_gas_meter().unwrap();
+        gas_meter
             .charge_linear_gas(
                 &<S as GasSpec>::gas_to_charge_per_evm_gas(),
                 gas_used as u32,
             )
-            .unwrap();
-        let gas_meter = state.try_as_basic_gas_meter().unwrap();
+            .expect("No underflow is possible here as we init EVM gas with gas meter gas");
         let total_gas_used =
             gas_meter.initial_gas.as_ref()[0] - gas_meter.remaining_gas.as_ref()[0];
-        const RELATIVE_MARGIN: u64 = 100_000;
-        let gas_used_with_margins = (total_gas_used * 3) / 2 + RELATIVE_MARGIN; // gas * 1.5 + 100_000
-        Ok(U64::from(gas_used_with_margins))
+        Ok(U64::from(apply_margins(total_gas_used)?))
     }
 
     /// Handler for: `debug_traceTransaction`
@@ -414,18 +411,13 @@ where
                 continue;
             }
 
-            transact_commit(
-                &mut evm_db,
-                block_env.clone(),
-                replay_tx_env(&tx),
-                cfg_env.clone(),
-            )
-            .map_err(|e| eth_api_into_rpc_error(eth_from_evm_error(e)))?;
+            transact_commit(&mut evm_db, &block_env, replay_tx_env(&tx), cfg_env.clone())
+                .map_err(|e| eth_api_into_rpc_error(eth_from_evm_error(e)))?;
         }
 
         // Trace the target transaction
         self.trace_transaction(
-            block_env.clone(),
+            block_env,
             replay_tx_env(&traced_tx),
             cfg_env,
             evm_db,
@@ -444,6 +436,15 @@ pub enum PendingOrBlock {
     Number(u64),
     /// Invalid block number.
     Invalid(String),
+}
+
+const ABSOLUTE_MARGIN: u64 = 100_000;
+/// gas * 1.5 + 100_000
+fn apply_margins(gas: u64) -> Result<u64, RpcInvalidTransactionError> {
+    (gas / 2)
+        .checked_mul(3)
+        .and_then(|with_relative_margin| with_relative_margin.checked_add(ABSOLUTE_MARGIN))
+        .ok_or(RpcInvalidTransactionError::GasUintOverflow)
 }
 
 impl<S: Spec> Evm<S>
@@ -476,7 +477,7 @@ where
                     let mut inspector = TracingInspector::new(inspector_config);
 
                     let gas_limit = tx_env.gas_limit;
-                    let res = inspect(db, block_env, tx_env, cfg, &mut inspector)?;
+                    let res = inspect(db, &block_env, tx_env, cfg, &mut inspector)?;
                     inspector.set_transaction_gas_limit(gas_limit);
 
                     let frame = inspector
@@ -504,7 +505,7 @@ where
         let cfg_env = get_cfg_env(&block_env, cfg, Some(get_cfg_env_template()));
         let evm_db: EvmDb<_, S> = self.get_db(state);
 
-        executor::call(evm_db, block_env, tx_env, cfg_env)
+        executor::call(evm_db, &block_env, tx_env, cfg_env)
             .map_err(|err| eth_api_into_rpc_error(eth_from_evm_error(err)))
     }
 
@@ -513,14 +514,18 @@ where
         &self,
         block_number: u64,
         state: &mut ApiStateAccessor<S>,
-    ) -> MaybeSealedBlock {
+    ) -> Option<MaybeSealedBlock> {
         let block = self.blocks.get(&block_number, state).unwrap_infallible();
         if let Some(block) = block {
-            return MaybeSealedBlock::Sealed(block);
+            return Some(MaybeSealedBlock::Sealed(block));
         }
 
         let pending = self.pending_block(state);
-        MaybeSealedBlock::Pending(pending)
+        if block_number == pending.header.number {
+            return Some(MaybeSealedBlock::Pending(pending));
+        }
+
+        None
     }
 
     /// Convert string to block nr.
@@ -566,14 +571,14 @@ where
         &self,
         block_number: Option<String>,
         state: &mut ApiStateAccessor<S>,
-    ) -> Option<SealedBlock> {
+    ) -> Option<MaybeSealedBlock> {
         let pending_or_block_nr = self.str_to_block_nr(block_number, state);
 
         match pending_or_block_nr {
-            PendingOrBlock::Number(nr) => self.blocks.get(&nr, state).unwrap_infallible(),
+            PendingOrBlock::Number(nr) => self.get_maybe_sealed_block(nr, state),
             PendingOrBlock::Pending => {
                 let pending_block = self.pending_block(state);
-                Some(pending_block.seal())
+                Some(MaybeSealedBlock::Pending(pending_block))
             }
             PendingOrBlock::Invalid(invalid) => {
                 tracing::error!(invalid, "Invalid block number");
@@ -620,6 +625,11 @@ where
                 .timestamp
                 .try_into()
                 .expect("The impossible happened: timestamp overflow u64"),
+            excess_blob_gas: current_block_env
+                .blob_excess_gas_and_price
+                .map(|blob_gas| blob_gas.excess_blob_gas),
+            base_fee_per_gas: Some(current_block_env.basefee),
+
             ..Default::default()
         };
 
@@ -661,18 +671,18 @@ where
         block_number: Option<String>,
         state: &mut ApiStateAccessor<S>,
     ) -> Result<BlockEnv, EthApiError> {
-        let block = match block_number {
-            Some(ref block_number) if block_number == "pending" => {
-                self.block_env.get(state).unwrap_infallible()
-            }
-            _ => self
-                .get_sealed_block_by_number(block_number, state)
-                .map(BlockEnv::from),
-        };
-        let Some(block) = block else {
-            return Err(EthApiError::UnknownBlockOrTxIndex);
-        };
-        Ok(block)
+        let maybe_blcok = self
+            .get_sealed_block_by_number(block_number, state)
+            .ok_or(EthApiError::UnknownBlockOrTxIndex)?;
+
+        Ok(match maybe_blcok {
+            MaybeSealedBlock::Pending(_) => self
+                .block_env
+                .get(state)
+                .unwrap_infallible()
+                .expect("The impossible happened: block_env is not set."),
+            MaybeSealedBlock::Sealed(sealed_block) => BlockEnv::from(sealed_block),
+        })
     }
 }
 
@@ -717,7 +727,7 @@ fn get_cfg_env_template() -> CfgEnv {
 // modified from: https://github.com/paradigmxyz/reth many times
 pub(crate) fn build_rpc_receipt(
     block: MaybeSealedBlock,
-    tx: TransactionSignedAndRecovered,
+    tx: TxSignedAndRecovered,
     tx_number: u64,
     receipt: Receipt,
 ) -> TransactionReceipt {