+++ crates/module-system/sov-modules-api/src/state/accessors/checkpoints.rs
@@ -7,6 +7,7 @@ use super::internals::{AccessoryDelta, Delta};
 use super::temp_cache::{CacheLookup, TempCache};
 use super::{BootstrapWorkingSet, BorshSerializedSize, UniversalStateAccessor};
 use crate::capabilities::{Kernel, RollupHeight};
+use crate::state::accessors::internals::FirstTimeReads;
 use crate::state::traits::PerBlockCache;
 #[cfg(feature = "native")]
 use crate::TxChangeSet;
@@ -49,6 +50,11 @@ impl<S: Spec> StateCheckpoint<S> {
     /// Check if key is in the cache.
     pub fn is_value_cached(&self, namespace: Namespace, key: &SlotKey) -> IsValueCached {
         self.delta.is_value_cached(namespace, key)
+    }
+
+    /// Keys and values that were read for the first time.
+    pub fn first_reads(&self) -> FirstTimeReads {
+        self.delta.first_reads()
     }
 
     /// Commits the revertable part of the `StateCheckpoint` cache.
@@ -224,7 +230,7 @@ impl<S: Spec> StateCheckpoint<S> {
     // TODO: Remove this method if we stop using `StateCheckpoint` in the sequencer
     #[cfg(feature = "native")]
     pub fn apply_tx_changes(&mut self, changeset: TxChangeSet) {
-        for ((key, namespace), value) in changeset.0 {
+        for ((key, namespace), value) in changeset.writes {
             if let Some(value) = value {
                 self.set_value(namespace, &key, value);
             } else {