+++ crates/module-system/sov-solana-offchain-auth/src/authentication.rs
@@ -83,7 +83,7 @@ pub struct SolanaOffchainSimpleMessage<S: Spec> {
 pub const PREAMBLE_LEN: usize = 85;
 
 /// The preamble/header required for signing solana offchain messages, supporting a single signer.
-/// See https://docs.anza.xyz/proposals/off-chain-message-signing#message-preamble
+/// See <https://docs.anza.xyz/proposals/off-chain-message-signing#message-preamble>
 #[derive(BorshSerialize, BorshDeserialize)]
 pub struct RawSolanaOffchainMessagePreamble {
     pub signing_domain: [u8; 16],