+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2025-09-25
+- #1758 A new config value in rollup.toml: `runner.da_total_timeout_secs`. It should be larger than the total time da service attempts to fetch data. Default value is 10 minutes.
+
+# 2025-09-21
+- #1729 Adds optional `num_cache_warmup_workers` to `xxx_rollup_config.toml`. This field specifies the number of workers responsible for warming up the executor cache in the preferred sequencer.
+- #1752 Changes in benchmark for sov-demo-rollup
+
 # 2025-09-19
 - #1723 **Breaking change** Introduced an optional `[sequencer.extension]` section in the rollup config, this section is required for `EVM` rollups.
 - #1718 Implement support for historical state queries in ethereum RPC methods.